moved shared folder inside react app so it could be bundled
