+++ backend/src/db/dynamo/DynamoTrainerRequestDao.ts
@@ -1,4 +1,4 @@
-import TrainerClientPair from "../../../../shared/models/TrainerClientPair";
+import TrainerClientPair from "../../../../react-native-app/shared/models/TrainerClientPair";
 import TrainerRequestDao from "../TrainerRequestDao";
 import TableAccessObject from "./TableAccessObject";
 import DB_TABLES from "./tables";
