+++ backend/src/db/dynamo/DynamoUserDao.ts
@@ -1,4 +1,4 @@
-import User from "../../../../shared/models/User";
+import User from "../../../../react-native-app/shared/models/User";
 import UserDao from "../UserDao";
 import TableAccessObject from "./TableAccessObject";
 import DB_TABLES from "./tables";
