+++ backend/src/handlers/update-user.ts
@@ -1,0 +1,20 @@
+import { APIGatewayProxyEventV2 } from "aws-lambda";
+import LambaUtils from "../utils/LambdaUtils";
+import UserService from "../services/UserService";
+import User from "../../../react-native-app/shared/models/User";
+
+export const handler = async (event: APIGatewayProxyEventV2) => {
+    return await LambaUtils.handleEventWithBody<User>(event, async (newUserData,res)=>{
+        try {
+            await new UserService().updateUser(newUserData)
+            res.setBodyToMessage("Successfully updated user!");
+            res.setCode(200);
+        }
+        catch (e) {
+            console.log("Error updating user:", e);
+            res.setBodyToMessage("Could not update user");
+            res.setCode(500);
+        }
+        return res;
+    });
+}
