--- react-native-app/models/User.ts
@@ -1,11 +1,0 @@
-export default class User {
-    public id: string;
-    public firstname: string;
-    public lastname: string;
-
-    constructor(id:string,firstname:string,lastname:string) {
-        this.id = id;
-        this.firstname = firstname;
-        this.lastname = lastname;
-    }
-}
