+++ react-native-app/shared/constants/CoordinateConstants.ts
