+++ react-native-app/shared/models/CoordinateModels.ts
