+++ react-native-app/shared/models/LocationCheck.ts
