+++ react-native-app/shared/models/RestaurantData.ts
