+++ react-native-app/shared/models/TrainerClientPair.ts
