+++ react-native-app/shared/models/User.ts
@@ -1,0 +1,10 @@
+export default class User {
+    constructor(
+        public userId: string,
+        public email: string,
+        public firstname?: string,
+        public lastname?: string,
+        public phone?: string,
+    ) {}
+}
+
