+++ react-native-app/views/UserProfileView.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Button, StyleSheet, Text, View } from 'react-native';
 import Profile from '../components/Profile';
-import User from '../models/User';
+import User from '../shared/models/User';
 
 const dummyUser = new User("userId","Andre","Miller");
