--- shared/models/Response.ts
@@ -1,6 +1,0 @@
-export default class Response<TData> {
-    public data?: TData;
-    public code?: number;
-    public message?: string;
-    public success?: boolean;
-}
